requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.35
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
python-dotenv==1.0.1
Flask-Login==0.6.3
httpx==0.27.2
requests==2.32.3
celery==5.4.0
redis==5.0.8
rapidfuzz==3.9.6
python-whois==0.9.5
dnspython==2.6.1
email-validator==2.2.0